Reorganize cachito to use the web and workers subpackages
