+++ .travis.yml
@@ -2,7 +2,7 @@ language: python
 python:
   - "3.6"
 install:
-  - pip install -r dev-requirements.txt
+  - pip install -r requirements-dev.txt
   - pip install tox
 script:
   - tox
