+++ cachito/web/wsgi.py
@@ -1,5 +1,5 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
-from cachito.app import create_app
+from cachito.web.app import create_app
 
 
 app = create_app()
