+++ requirements-workers.txt
@@ -1,0 +1,1 @@
+# TODO: Add workers requirements here
