+++ requirements.txt
@@ -1,3 +1,1 @@
-docker
 flask
-requests
