+++ tox.ini
@@ -3,7 +3,7 @@ envlist = lint,py36
 
 [testenv]
 deps =
-    -rdev-requirements.txt
+    -rrequirements-dev.txt
 whitelist_externals =
     make
 setenv =
